주기 변경 @RequestBody 없어서 넘어오는 값 null인 문제 해결

+++ src/main/java/com/example/plantServer/entity/Pot.java
@@ -25,15 +25,13 @@ public class Pot {
     private String serialId;
 
     @Column(name="humidity")
-    private Float humidity;
-//    private Integer humidity;
+    private Integer humidity;
 
     @Column(name="soil_humidity")
     private Integer soil_humidity;
 
     @Column(name="temper")
-    private Float temper;
-//    private Integer temper;
+    private Integer temper;
 
     @Column(name="water_level")
     private Integer waterLevel;
